requirements.txt
python-telegram-bot==20.7
numpy==1.24.3
python-dotenv==1.0.0
